Interactive Card Details Form
Building out an interactive card details form using  semantic HTML .

The layout of the page is designed using CSS Grid and Flexbox. Media queries are used to ensure the page works welll on mobile / tablet / desktop.

Javascript is used to validate the form inputs to ensure they're not blank and that they use the correct format for each corresponding field.